+++ pulse2percept/implants/retina/imie.py
@@ -59,14 +59,8 @@ class IMIE(RetinalImplant):
             self.shape, e_spacing, z=z,
             electrode_type=DiskElectrode, radius=elec_radius)
         
-        # Set left/right eye:
-        if not isinstance(eye, str):
-            raise TypeError("'eye' must be a string, either 'LE' or 'RE'.")
-        if eye != 'LE' and eye != 'RE':
-            raise ValueError("'eye' must be either 'LE' or 'RE'.")
-        self.eye = eye
         # Unfortunately, in the left eye the labeling of columns is reversed...
-        if eye == 'LE':
+        if self.eye == 'LE':
             # TODO: Would be better to have more flexibility in the naming
             # convention. This is a quick-and-dirty fix:
             names = self.electrode_array.electrode_names